hw_10/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.6
dnspython==2.6.1
mongoengine==0.28.2
psycopg2==2.9.9
pymongo==4.7.2
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1